Salt + Argon2i Hashing
NEW DLL REQUIREMENT: Source/argon2/lib/Argon2RefDll.dll - it is possible
to use Argon without this DLL . . but it's easier to include it this way
. . sorry

Sha512 (without a salt no less) is considered unsecure.

So. . I switched it to Argon2i Hashing
Link: https://github.com/P-H-C/phc-winner-argon2

Several settings can be changed via the Encryption.h and Encryption.cpp
files.
If you change the length of the salt and/or hash, make sure to update
the sql file accordingly:

`password` varchar ( preHashSize * 2 )
`salt` varchar ( saltSize )    Note: saltSize and preHashSize found in
Encryption.h